# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an API for the Web App to be developed that builds on the IMDB Movies Dataset available at Kaggle.",
    "version": "1.0.0",
    "title": "Web App"
  },
  "host": "localhost:3000",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/movies/id/{id}": {
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "All the available information about a movie found by the ID of the movie",
        "description": "",
        "operationId": "getInformationMoviesID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the Movie",
            "required": true,
            "type": "integer"
          },
          {
            "name": "contentType",
            "in": "query",
            "description": "The Content Type of the Response (JSON/CSV)",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "rating": {
                  "type": "string"
                },
                "year": {
                  "type": "integer",
                  "format": "int32"
                },
                "users_rating": {
                  "type": "string"
                },
                "votes": {
                  "type": "string"
                },
                "metascore": {
                  "type": "string"
                },
                "img_url": {
                  "type": "string"
                },
                "countries": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "languages": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "actors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "genre": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "tagline": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "directors": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "runtime": {
                  "type": "string"
                },
                "imdb_url": {
                  "type": "string"
                },
                "id": {
                  "type": "integer",
                  "format": "int32"
                },
                "links": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "rel": {
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      },
                      "href": {
                        "type": "string",
                        "format": "uri"
                      }
                    }
                  }
                }
              },
              "example": {
                "title": "A Rainy Day in New York",
                "rating": "PG-13",
                "year": 2019,
                "users_rating": "6.6",
                "votes": "21,903",
                "metascore": "44.0",
                "img_url": "https://m.media-amazon.com/images/M/MV5BODAwZDlhZjUtYzM2MS00MGVmLWFjNWMtODc5NjM2OTNkNjExXkEyXkFqcGdeQXVyMzIzNDU1NTY@._V1_UX182_CR0,0,182,268_AL__QL50.jpg",
                "countries": "['USA']",
                "languages": "['English']",
                "actors": "['Timothée Chalamet', 'Elle Fanning', 'Liev Schreiber', 'Suzanne Smith', 'Olivia Boreham-Wing', 'Ben Warheit', 'Griffin Newman', 'Selena Gomez', 'Gus Birney', 'Elijah Boothe', 'Will Rogers', 'Annaleigh Ashford', 'Jude Law', 'Frank Marzullo', 'Kirby Mitchell']",
                "genre": "['Comedy', 'Romance']",
                "tagline": "",
                "description": "A young couple arrives in New York for a weekend where they are met with bad weather and a series of adventures.",
                "directors": "['Woody Allen']",
                "runtime": "92 min",
                "imdb_url": "https://www.imdb.com/title/tt7139936/",
                "id": 1,
                "links": [
                  {
                    "rel": "self",
                    "method": "GET",
                    "href": "/api/movies/id/1"
                  },
                  {
                    "rel": "omdb",
                    "method": "GET",
                    "href": "/api/omdb/A Rainy Day in New York"
                  }
                ]
              }
            },
            "headers": {
              "title": {
                "type": "string",
                "description": "The Title of the Movie"
              },
              "rating": {
                "type": "string",
                "description": "The Rating of the Movie"
              },
              "year": {
                "type": "integer",
                "format": "int32",
                "description": "The Release Year of the Movie"
              },
              "users_rating": {
                "type": "string",
                "description": "The Users Rating of the Movie"
              },
              "votes": {
                "type": "string",
                "description": "The Votes on the Movie"
              },
              "metascore": {
                "type": "string",
                "description": "The Metascore of the Movie"
              },
              "img_url": {
                "type": "string",
                "description": "The URL of the Image"
              },
              "countries": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "The Countries the Movie was Filmed"
              },
              "languages": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "The Languages the Movie is Available"
              },
              "actors": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "The Actors in the Movie"
              },
              "genre": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "The Genres of the Movie"
              },
              "tagline": {
                "type": "string",
                "description": "The Tagline of the Movie"
              },
              "description": {
                "type": "string",
                "description": "The Description of the Movie"
              },
              "directors": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "The Directors of the Movie"
              },
              "runtime": {
                "type": "string",
                "description": "The Runtime of the Movie"
              },
              "imdb_url": {
                "type": "string",
                "description": "The URL of the IMDB Page of the Movie"
              },
              "id": {
                "type": "integer",
                "description": "The ID of the Movie"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/api/actors/id/{id}": {
      "get": {
        "tags": [
          "actors"
        ],
        "summary": "Get an actor by the ID",
        "description": "",
        "operationId": "getActorWithID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the Actor",
            "required": true,
            "type": "integer"
          },
          {
            "name": "contentType",
            "in": "query",
            "description": "The Content Type of the Response (JSON/CSV)",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "id": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "links": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "rel": {
                          "type": "string"
                        },
                        "method": {
                          "type": "string"
                        },
                        "href": {
                          "type": "string",
                          "format": "uri"
                        }
                      }
                    }
                  }
                }
              },
              "example": [
                {
                  "name": "$2 Tony",
                  "id": 1,
                  "links": [
                    {
                      "rel": "see",
                      "method": "GET",
                      "href": "/api/actors/id/1"
                    }
                  ]
                }
              ]
            },
            "headers": {
              "name": {
                "type": "string",
                "description": "The Name of the Actor"
              },
              "id": {
                "type": "integer",
                "format": "int32",
                "description": "The ID of the Actor"
              },
              "links": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "The Parts of the HTTP Request"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}